Kinda use callbacks.
Rather than assigning correlation ids to outgoing requests, or maintaing
a queue of outstanding requests per peer in the model, instead we pass a
function from the response to an action to the network handler, and the
network handler will send us back the result of applying the function to the
response.

It's kinda horrible, but it works. For now.